refactor DOCX conversion: Split writer functionality into separate module, enhance image processing, and clean up utility functions
